+++ tests/buffer/gaussian_transform.rs
@@ -133,7 +133,7 @@ fn test_gaussian_transform_buffer_update_should_update_buffer_correctly() {
 }
 
 #[test]
-fn test_gaussian_transform_buffer_try_from_and_into_wgpu_buffer_should_be_equal() {
+fn test_gaussian_transform_buffer_wgpu_buffer_round_trip_should_be_equal() {
     let ctx = TestContext::new();
     let pod = GaussianTransformPod::new(
         1.0,