+++ tests/common/given.rs
@@ -56,12 +56,12 @@ pub fn gaussian_with_seed(seed: u32) -> Gaussian {
 
     let pos = Vec3::new(base + 1.1, base + 2.2, base + 3.3);
 
-    let color = U8Vec4::new(
-        ((base + 11.0) % 256.0) as u8,
-        ((base + 22.0) % 256.0) as u8,
-        ((base + 33.0) % 256.0) as u8,
-        ((base + 44.0) % 256.0) as u8,
-    );
+    let color = Vec4::new(
+        (base + 11.0) % 256.0,
+        (base + 22.0) % 256.0,
+        (base + 33.0) % 256.0,
+        (base + 44.0) % 256.0,
+    ) / 255.0;
 
     let mut sh = [Vec3::ZERO; 15];
     for (i, sh) in sh.iter_mut().enumerate() {